requirements.txt
# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.25.2

# Email
email-validator==2.1.0

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2
pytz==2023.3

# Logging and CLI
colorama==0.4.6
rich==13.7.0

